API Token enhancement
* splits view render and token generating in two different functions
* adds access to api root in template

+++ api/views.py
@@ -1,7 +1,7 @@
 # Create your views here.
 from django.core.exceptions import ObjectDoesNotExist
 from django.db.models import Q
-from django.http import HttpRequest
+from django.http import HttpRequest, HttpResponse
 from django.shortcuts import render, redirect
 from django.urls import reverse
 from django.utils.decorators import method_decorator
@@ -24,11 +24,7 @@ from service.models import Service, Layer, Metadata
 from structure.models import Organization, MrMapGroup, Role, Permission
 from users.helper import user_helper
 
-@check_permission(
-    Permission(
-        can_generate_api_token=True
-    )
-)
+
 def menu_view(request: HttpRequest):
     """ The API menu view where settings for the remote access can be set.
 
@@ -50,30 +46,63 @@ def menu_view(request: HttpRequest):
         # User has no token yet
         token = None
 
+    if token is not None:
+        token_form = TokenForm(instance=token)
+
+    token_form.action_url = reverse("api:generate-token")
+    params = {
+        "form": token_form,
+    }
+    default_context = DefaultContext(request, params)
+    return render(request, template, default_context.get_context())
+
+
+@check_permission(
+    Permission(
+        can_generate_api_token=True
+    )
+)
+def generate_token(request: HttpRequest):
+    """ Generates a token for the user.
+
+    Returns after finished work back to the menu page
+
+    Args:
+        request (HttpRequest): The incoming request
+    Returns:
+         A redirect to a view
+    """
     if request.method == "POST":
+        user = user_helper.get_user(request)
+        # Get user token
+        try:
+            token = Token.objects.get(
+                user=user
+            )
+        except ObjectDoesNotExist:
+            # User has no token yet
+            token = None
+
+        # Remove the given key.
+        # Otherwise the is_valid() would fail, since the key for this user could already exists.
+        # We are only interested in the csrf token validation.
+        post_dict = request.POST.dict()
+        post_dict["key"] = ""
+        token_form = TokenForm(post_dict)
+
+        if not token_form.is_valid():
+            return HttpResponse(status=403)
+
         # Generate new access token, old token can not be reused, must be deleted
         if token is not None:
             token.delete()
         token = Token(user=user)
         token.save()
 
-        # Redirect user directly to the same page. Why?
-        # This way, we make sure the user does not re-generate another token accidentally by pressing F5 or reload,
-        # or whatever. We force the GET way.
-        return redirect("api:menu")
-    else:
-        # Show existing access token
-        pass
-
-    if token is not None:
-        token_form = TokenForm(instance=token)
-
-    token_form.action_url = reverse("api:menu")
-    params = {
-        "form": token_form,
-    }
-    default_context = DefaultContext(request, params)
-    return render(request, template, default_context.get_context())
+    # Redirect user directly to the same page. Why?
+    # This way, we make sure the user does not re-generate another token accidentally by pressing F5 or reload,
+    # or whatever. We force the GET way.
+    return redirect("api:menu")
 
 
 class APIPagination(PageNumberPagination):
